support row based replication, read and write of MySQL 5.7.x JSON binary
Oracle MySQL stores JSON as compact binary,  MariaDB stores JSON as
original text, as noted by https://mariadb.com/kb/en/library/json-data-type/,
MariaDB doesn't support row based replication of JSON binary from Oracle
MySQL, also can't directly access MySQL table including JSON binary column.

This patch fixed these issues, and as a bonus(or hack), it supports
updating JSON binary column in place although the modified table can't
be recognized by Oracle MySQL anymore because MariaDB actually writes JSON
text instead of JSON binary.

+++ plugin/auth_pam/testing/pam_mariadb_mtr.c
@@ -56,7 +56,7 @@ int pam_sm_authenticate(pam_handle_t *pamh, int flags,
     if (strcmp(r1, "crash pam module") == 0 && atoi(r2) == 616)
       abort();
 
-    if (strlen(r1) == (uint)atoi(r2) % 100)
+    if (strlen(r1) == (size_t)atoi(r2) % 100)
       retval = PAM_SUCCESS;
     else
       retval = PAM_AUTH_ERR;
